[subscript-experiment] consolidate duplicate code
